Extract tests for isArray to own file
